Move signup form test util functions to separate file
